refactor: divide the persona select panel component
