+++ packages/mask/src/components/shared/ConnectPersonaSelect/PersonaSelectPanelDialog.tsx
@@ -1,0 +1,100 @@
+import { DialogContent } from '@mui/material'
+import { InjectedDialog } from '@masknet/shared'
+import { useState } from 'react'
+import { useI18N } from '../../../utils'
+import { useRemoteControlledDialog } from '@masknet/shared-base-ui'
+import { makeStyles } from '@masknet/theme'
+import { PluginNextIDMessages } from '../../../plugins/NextID/messages'
+import { PersonaSelectPanel, PositionOption } from './PersonaSelectPanel'
+
+type PositionStyle = {
+    top?: number
+    right?: number
+    position?: 'absolute'
+}
+
+const useStyles = makeStyles<{
+    positionStyle: {
+        top?: number
+        right?: number
+        position?: 'absolute'
+    }
+}>()((theme, props) => {
+    return {
+        root: {
+            width: 384,
+            height: 386,
+            padding: theme.spacing(1),
+            background: theme.palette.maskColor.bottom,
+            position: props.positionStyle.position,
+            top: props.positionStyle.top,
+            right: props.positionStyle.right,
+        },
+        content: {
+            padding: theme.spacing(0, 2, 2, 2),
+            scrollbarWidth: 'none',
+            '&::-webkit-scrollbar': {
+                display: 'none',
+            },
+        },
+        header: {
+            background: `${theme.palette.maskColor.bottom} !important`,
+        },
+        items: {
+            overflow: 'auto',
+            scrollbarWidth: 'none',
+            '&::-webkit-scrollbar': {
+                display: 'none',
+            },
+        },
+    }
+})
+
+const positionStyleMap: Record<PositionOption, PositionStyle> = {
+    center: {},
+    'top-right': {
+        position: 'absolute',
+        top: 0,
+        right: 0,
+    },
+}
+
+export const PersonaSelectPanelDialog = () => {
+    const { t } = useI18N()
+
+    const [finishTarget, setFinishTarget] = useState<string>()
+    const [position, setPosition] = useState<PositionOption>('center')
+    const [enableVerify, setEnableVerify] = useState(true)
+    const { classes } = useStyles({ positionStyle: positionStyleMap[position] })
+
+    const { open, closeDialog } = useRemoteControlledDialog(
+        PluginNextIDMessages.PersonaSelectPanelDialogUpdated,
+        (ev) => {
+            if (!ev.open) {
+                setFinishTarget(undefined)
+            } else {
+                setFinishTarget(ev.target)
+                setEnableVerify(ev.enableVerify)
+                setPosition(ev.position ?? 'center')
+            }
+        },
+    )
+
+    return open ? (
+        <InjectedDialog
+            disableTitleBorder
+            open={open}
+            classes={{
+                paper: classes.root,
+                dialogTitle: classes.header,
+            }}
+            maxWidth="sm"
+            onClose={closeDialog}
+            title={t('applications_persona_title')}
+            titleBarIconStyle="close">
+            <DialogContent classes={{ root: classes.content }}>
+                <PersonaSelectPanel enableVerify={enableVerify} finishTarget={finishTarget} onClose={closeDialog} />
+            </DialogContent>
+        </InjectedDialog>
+    ) : null
+}
